requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Django==2.0
django-ffmpeg==0.0.11
django-haystack==2.8.1
django-videokit==0.0.4
django-widget-tweaks==1.4.3
idna==2.7
kombu==4.2.1
Pillow==5.3.0
pytz==2018.5
requests==2.20.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.24
vine==1.1.4
